services/ingestion/requirements.txt
kafka-python>=2.0.2
websockets>=11.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
colorlog>=6.7.0
